Overhaul the encounter handling process
While fairly functionally similar to what was already in place,
this gives us the ability to better detect when we end encounters
prematurely, and puts us in a better position to handle other
types of fights -- namely, Mythic+ dungeons.

Unfortunately, in order to retain support for compressed logfiles,
this means we need to abandon the easier-to-use bufio.Scanner, in
favour of the more raw bufio.Reader. This is purely so we can better
handle those encounters that end prematurely, as the scanner
has no Seek or Peek functionality.

(Ideally, we don't pass a bufio object at all, and just pass the
filename around so we can Seek, but that was considerably more
invasive than what we have here.)